fix(testsuite): fix testsuite launch, start modes and add of second test stage depending on test type

+++ tests/testsuite.py
@@ -1,37 +1,58 @@
 import os
+
 import pytest
 import subprocess as sp
-import time
 
-folders = ["bind", "minimal", "good", "syntax", "type"]
+folders = ["tests/bind", "tests/minimal", "tests/good", "tests/syntax", "tests/type"]
 param_filepaths = []
 for folder in folders:
-    param_filepaths += [(open(f"{folder}/{file}", 'r').name, f"{folder}/{file}") for file in os.listdir(folder)]
+    param_filepaths += [(open(f"{folder}/{file}", 'r').name, folder.split("/")[-1]) for file in os.listdir(folder)]
 
+def get_ref_return_code(test_type: str) -> int:
+    if test_type == "minimal" or test_type == "good":
+        return 0
+    elif test_type == "syntax":
+        return 3
+    elif test_type == "bind":
+        return 4
+    elif test_type == "type":
+        return 5
+    return 0
 
-@pytest.mark.parametrize("name,filepath", param_filepaths)
-def test_file(name, filepath):
-    executable = "../src/tc"
+def get_mode(test_type : str) -> str:
+    if test_type == "minimal" or test_type == "good":
+        return "-bBA"
+    elif test_type == "syntax":
+        return "-A"
+    elif test_type == "bind":
+        return "-bBA"
+    elif test_type == "type":
+        return "-beBEA"
+    return "-A"
 
-    proc = sp.Popen([executable, "-A", filepath], stdout=sp.PIPE, stderr=sp.STDOUT, bufsize=0)
+@pytest.mark.parametrize("filepath,test_type", param_filepaths)
+def test_file(filepath, test_type):
+    executable = "./src/tc"
+
+    proc = sp.Popen([executable, get_mode(test_type), "--parse", filepath], stdout=sp.PIPE, stderr=sp.STDOUT, bufsize=0)
     try:
         out, err = proc.communicate(timeout=0.5)
         assert err is None
         assert out is not None
-        assert proc.returncode == 0
+        assert proc.returncode == get_ref_return_code(test_type)
 
-        proc_piped = sp.Popen([executable, "-"], stdin=sp.PIPE, stdout=sp.PIPE, stderr=sp.STDOUT, bufsize=0)
+        if test_type == "minimal" or test_type == "good":
+            proc_piped = sp.Popen([executable, "-"], stdin=sp.PIPE, stderr=sp.STDOUT, bufsize=0)
 
-        proc_piped.stdin.write(out)
-        proc_piped.stdin.close()
+            proc_piped.stdin.write(out)
+            proc_piped.stdin.close()
 
-        try:
-            out_piped, err_piped = proc.communicate(timeout=0.5)
-            assert out_piped is not None
-            assert err_piped is None
-            assert proc_piped.returncode == 0
-        finally:
-            proc_piped.kill()
-
+            try:
+                out_piped, err_piped = proc_piped.communicate()
+                assert out_piped is None
+                assert err_piped is None
+                assert proc_piped.returncode == 0
+            finally:
+                proc_piped.kill()
     finally:
         proc.kill()
